backend/requirements.txt
fastapi
python-multipart
aiofiles
streamz
openai
elevenlabs
pydantic
uvicorn
python-dotenv